Updated openstack/openstack Project: openstack/cinder  aed5591aeaf4aaecf22a1efba6f059dc1977f79f
Support A/A in delete actions and get_capabilities

This patch allows deleting operations and getting capabilities in HA
Active-Active configurations.

To allow this we have to use the cluster topic queue in the Message
Broker, so following RPC methods in cinder/volume/rpcapi have been
changed:

- delete_volume
- delete_snapshot
- delete_consistencygroup
- delete_cgsnapshot
- delete_group
- delete_group_snapshot

When the cluster field is not set all RPC calls will behave as before,
sending it to the host topic queue.

Change-Id: I8bcf26afec7f2c54f281dc1ac771825b1c1625ad
Specs: https://review.openstack.org/327283
Implements: blueprint cinder-volume-active-active-support