Updated openstack/openstack Project: openstack/ironic  d217b5cf4f9c16ce5909c73df54f58dd01162c19
Add release names & numbers to API version history

This should make it easier to work out when an API change was
introduced and what API microversions are likely to be supported by a
given cloud. Release numbers were determined by reading the release
notes in most cases, and by searching Git history when that failed. I
was unable to determine when the API versions 1.1-1.6 were released, so
have listed them all as simply "Kilo".

Keeping this list up-to-date is (currently?) a manual process; this
patch adds "Update the API version history" to the how-to-release
document as a pre-release task.

Change-Id: I44b29d29e38b0ac1b256c2377ff84db2482b7aa7