Updated openstack/openstack Project: openstack-infra/project-config  853df3ccaa773fa4396ea21e911a790e0b7ae3a7
Migrate monasca-analytics unittests to Xenial

This change is a little tricky because we have to make sure we don't
conflict with the job defs in the python-jobs group. So we remove the
python-jobs group from this project, redefine the portions we want, then
sprinkle in the special sauce unittest job.

Changes to branches >= newton will now run this job on Xenial. Older
branches on Trusty.

Change-Id: I737ee23335fadff8e183c99c7e2e0389b900a7d5