Updated openstack/openstack Project: openstack-infra/project-config  b2ad55060dec1d33ec312a2c65c4975053ab5c4a
Simplify networking-odl JJB defs

This was part of the Xenial transition work. In order to understand of
this project had completed the transition I had to remove unused jobs,
compress existing job defs, and move comments around to where they are
supposed to be located just to understand what was going on.

I think this is much easier to understand now so pushing it up.

Change-Id: I6d9d3d343282f3b22892da7304623c9cc1f7387e