Updated openstack/openstack Project: openstack/zaqar-specs  86ef41ab9df84141039a01f616e6adbd24cd6849
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/8a0b36c22a945eb213ce1a739b62f0ed

Change-Id: I85cb18fe0bfe394aefbd37eda477334c7f5f3dc3