Updated openstack/openstack Project: openstack/python-zaqarclient  3e0d0ee79e7021e1515671c466c2f0aea38ba871
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/5c6ff379696d6744653a72a92c9b3bd8

Change-Id: Ia3d46379cdad604e7b2e0740b3fb57fbb67529c7