Updated openstack/openstack Project: openstack/python-ceilometerclient  158c97f75a1c3d672552f6164d09e6d0c0944098
Add __ne__ built-in function

In Python 3 __ne__ by default delegates to __eq__ and inverts the
result, but in Python 2 they urge you to define __ne__ when you
define __eq__ for it to work properly [1].There are no implied
relationships among the comparison operators. The truth of x==y
does not imply that x!=y is false. Accordingly, when defining __eq__(),
one should also define __ne__() so that the operators will behave as
expected.
[1]https://docs.python.org/2/reference/datamodel.html#object.__ne_

Change-Id: I4ba5b370e34cd64b13d87ef3ce2869d224dd0969