Updated openstack/openstack Project: openstack/python-novaclient  e490688b119786d8e59d6b72248d2484c032f761
Add min-disk and min-ram filters to list flavors

The Nova API allows user filter flavors by minDisk and minRam,
but that is not implemented in the current novaclient yet.
This patch introduce '--min-disk' and '--min-ram' options to support
it. In addition, We also append '--sort-key' and '--sort-dir' to
the flavor-list subcommand to support sort the result list.

Change-Id: I3849cc2d73057170dc224fa86b724f5141ca9a23
Closes-Bug: #1647867