Updated openstack/openstack Project: openstack-infra/project-config  cc7aa29a3a3d6ce05e097fb7d5bfaf81c4e41195
Switch to Xenial: oslo.messaging (part1):dsvm-functional

This is a first change for converting of oslo.messaging to Xenial with
the goal of running stable liberty/mitaka jobs on trusty and newer jobs
on Xenial.

For oslo.messaging-dsvm-functional change ostype to node parameter
instead.
Convert this job to Xenial, compress the job definition to better
understand it.

Move oslo.messaging-src-dsvm-full from devstack-gate.yaml to oslo.yaml.
Do not convert this to Xenial.

Remove unused {pipeline}-{injectedlib}-dsvm-{name}{job-suffix}
template.

Change-Id: I725c098d190d902bd34fb1791a9fd65fbc57fda9