Updated openstack/openstack Project: openstack/nova  2bb70e7b15e6cfef4652e2e49c4e02d151d2dbdf
Don't trace on ImageNotFound in delete_image_on_error

The point of the delete_image_on_error decorator is to
cleanup an image used during snapshot operations, so it
makes little sense to log an exception trace if the image
delete fails because the image no longer exists, which it
might not since _snapshot_instance method will proactively
delete non-active images in certain situations.

So let's just handle the ImageNotFound and ignore it.

Change-Id: I14e061a28678ad28e38bd185e3d0a35cae41a9cf
Closes-Bug: #1648574