Updated openstack/openstack Project: openstack-infra/project-config  f4ec81d33ff2cf968f29cbacce3e925a9edf39d7
Xenialify the grenade forward jobs

This transitions grenade forward jobs to run on Xenial for changes
against newton or master/ocata. The trusty vs xenial upgrade mapping is
the same as with normal grenade jobs except that we start this job on
the "previous" match so the branch we match against in zuul is
different.

Change-Id: Ia2cc1dbb212cfea7b6923384c676d0f6fa91dea0