Updated openstack/openstack Project: openstack/python-keystoneclient  7e893168283f757f34578a105cec0e65e1a0371d
Deprecate the generic client

The generic client has been around and unused for a really long time. I
dont think it works at all and was never updated to support V3 concepts.
Realistically we can probably just remove it and noone will notice, but
give it a quick deprecation cycle.

Closes-Bug: #1647930
Change-Id: Ie68c8995275bcd55aede49d8f4af4e0d172de089