Updated openstack/openstack Project: openstack-infra/project-config  11eb3a36419411b56230ed843e95bb51a4dfcffc
Change job-suffix to suffix

We have now some rather long job-template names, shorten them with
replacing job-suffix with suffix.

Before this change we had 1500 usages of job-suffix and 700 of just
suffix. This change therefore unifies to one setting.

Note that this does not change any job names at all, they are unchanged,
it changes and shortens only the names of job-templates.

Change-Id: I514df039a848d9f6619bca3adc05dfb84a761fc9