Updated openstack/openstack Project: openstack/ironic-python-agent  005927b80e6820ef7e3a24bf0389f7103429d0b5
Add HCTL to BlockDevices

This patch is adding a "hctl" attribute to the BlockDevices. HCTL stands
for: Host, Channel, Target and Lun, which is basically the SCSI address.

The idea behind this patch is to allow root device hints to find the
disk for deployment based on the SCSI address.

Partial-Bug: #1648036
Change-Id: If8897c68609e0df0378ee919b803ca5e497def02