Updated openstack/openstack Project: openstack/keystone  80c96b18e0aba931ca69e7f9b863eb5c4739f455
Expose idempotency issue with bootstrap

During some upgrade testing I was doing locally, I noticed an issue
where `keystone-manage bootstrap` isn't completely idempotent. This
is because `bootstrap` has the ability to recover lost admin accounts
by reseting the admin user's enabled status and updating their
password, regardless of it being different. This creates a revocation
event and causes admin tokens to be invalid after bootstrap is run
for a second time, making it not as idempotent as we'd like.

This commit introduces a test that exposes this behavior.

Change-Id: I627255b2b5d6ec401af2c07c4018930fea206e4a
Partial-Bug: 1647800