Updated openstack/openstack Project: openstack-infra/project-config  feac4e128fa87b8d213164de22f07b029058d0ae
Create repo for Enterprise Work Group Workload Reference Architectures.

The OpenStack Enterprise Work Group is developing a series of
'tenant-level' reference architectures to help users to learn and
understand how to deploy different types of workloads on an OpenStack
Cloud. These reference architectures describe the list of OpenStack
services that can be used to support the workload and provides sample code
to bootstrap the workload environment on an OpenStack Cloud.

This repo will be used to maintain the documents of these reference
architectures and sample code.

Change-Id: Icbb14771e577da74c8212ee6ce32b83ba648d8cd