Updated openstack/openstack Project: openstack/nova  f60b80931d8b1859d6b17092db1ad9983d37b59f
Replace uuid4() with uuidsentinel

As of now, in most of the test cases, uuidsentinel is used for
generating a UUID except at some places where uuid4() is used.
In order to maintain consistency, we propose to use uuidsentinel
module for generating UUIDs throughout the test cases.

There are some cases where unique UUIDs are required. For such
cases, generate_uuid() from oslo_utils.uuidutils is used.

Change-Id: Ie0e0fc1878e3f18065a11539b15d8c4ee893a29a