Updated openstack/openstack Project: openstack/nova  b494ea99798431e494c10aa6d71aa9e76961390f
Replace uuid4() with uuidsentinel

As of now, in most of the test cases, uuidsentinel is used for
generating a UUID except at some places where uuid4() is used.
In order to maintain consistency, we propose to use uuidsentinel
module for generating UUIDs throughout the test cases.

There are some cases where unique UUIDs are required. For such
cases, generate_uuid() from oslo_utils.uuidutils is used.

Change-Id: Ifaee2f79fc33d189751c4e3f405b579dd8d89784