Updated openstack/openstack Project: openstack/nova  4728c3e4fde5b5b7b068f60ea410d663deea7db2
libvirt: Fix initialising of LVM ephemeral disks

The LVM backend expects to write directly to the target disk rather
than to the image cache when initialising an ephemeral disk. This is
confounded by Image.cache(), which doesn't call the given callback
(_create_ephemeral in this case), if the target already exists.

Closes-Bug: #1648109
Co-authored-by: Feodor Tersin <ftersin@hotmail.com>
Change-Id: Ic9178d6c013670611e63d54a7669f35dc9770e91