Updated openstack/openstack Project: openstack/horizon  1ed9d3cd4f44212e6fdb44e1ac5a3e0d760a27e0
Remove type field and type description for create volume from a snapshot

The new volume should have the same type as the snapshot that the volume
is created from. Setting a different type will get an error from the API.

This behavior should be the same when creating a volume from the Volumes
tab and the Volume Snapshots tab.

Closes-bug: #1582458

Change-Id: I0443eb5c6cb18538ff76fc49d4018221691a2ea2