Updated openstack/openstack Project: openstack/nova  99b1a4a7b78ea50ff2a259980c7e4370729eeafc
Move get_instance() calls from try-except block

common.get_instance() raises a single exception HTTPNotFound only,
so it is not necessary to call the method in try-except block on
these cases.
This patch moves these calls from the block so that we will be
able to avoid digging what exceptions are raised on the method in
the future (On Ie1fc20fe5e028dd53116d5549814442232ddfe49 review,
we faced such situation).

Change-Id: I7c3898f79dfdd069c52de9327b7f5ec45786c8d6