Updated openstack/openstack Project: openstack/designate  93612c82972648276d53dd1f49b762a7b353ef29
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/3d9a86b5c5ffe39205b804601f6abc27

Change-Id: Ic0b3ff06362667de9f34be81e618726262706b92