Updated openstack/openstack Project: openstack/cinder  b5263b43d6ffe7acfe3eb6d9c48f44d942603db4
Admin API policy enforcement contingent on is_admin_project

In order for a user with the admin role to be able to perform
administrative actions, the role must be assigned to a project
that is deemed the "admin" project in the Keystone server. This
prevents someone being assigned admin on some random project
from being admin everywhere.

Change-Id: Ic4294cc1746702c345259c64bad1e20675a7d9ab
Closes-Bug: 968696