Updated openstack/openstack Project: openstack-dev/devstack  a2f7c21407cd20718ebb2bed8e364e522227a1e1
[opensuse] add python-xml to general deps list

the python-xml is a subpackage from the standard cpython package that
that contains elementtree and other bits that are needed almost
everywhere in OpenStack but isn't installed on a absolutely minimal
openSUSE Leap installation. This package doesn't exist on pip but
is a SUSE only invention, so just treat it similar to a bindep.

Change-Id: I82887c2e6895740d1b16d1269574519450ca783e