Updated openstack/openstack Project: openstack/nova  c6cb5cf1ba0b1483951f3d236c53ff7924f3d7af
Guestfs handle no passwd or group in image

When setting ownership of a file or directory, the guestfs driver
looks for the /etc/passwd and/or /etc/group files. In case they
are not found, the current driver lets the auges RuntimeError
through, which does not produce a very helpful error message.
Fixing that by handling the original exception and rasing a
Nova exception with more details in it.

Related-bug: #1646002

Change-Id: I2d15865c8be13b938e10e67c1b1b160f2a80f0c0