Updated openstack/openstack Project: openstack/nova  d22ef5bc587edab309cfa7eea4869c0768b8b45e
Fix race in test_volume_swap_server_with_error

This functional test is waiting for a flag to be set in
the CinderFixture which is called when swap_volume calls
volume_api.unreserve_volume after the swap_volume error
occurs. That happens after the two swap-volume notifications
are sent (for start and error).

The test was asserting a 5th notification for the generic
instance fault, which happens, but might not happen before
the CinderFixture flag is set and the test checks for 5
notifications.

This test really only cares that there are the two swap-volume
notifications, so to avoid a race this change just makes
the test assert that there are at least 4 versioned notifications
by the time the swap-volume error occurs, which is all it was
checking after that anyway.

Change-Id: I8fd44b083f109c034958a305a4e41f9d810c4389
Closes-Bug: #1639894