Updated openstack/openstack Project: openstack-infra/project-config  6f4d56065c7b3463af5c9f5b7b5686607c97f8dd
remove cueclient from osc check script

due to the current status of the cue project being independent,
no commits to the client in ~6 months [1], and no release in one
year [2]; i propose we remove the client from the osc checker.

it currently emits a deprecation warning, which was attempted to
be fixed [3] but abandoned since no one reviewed it.

[1] https://github.com/openstack/python-cueclient/commits/master
[2] https://github.com/openstack/python-cueclient/releases
[3] https://review.openstack.org/#/c/333747/

Change-Id: I6b5d3e74a227692e4c4f9ac1aac3af255d3726d2