Updated openstack/openstack Project: openstack/openstack-manuals  d48a6e0e1b244f1df1fbed525796341d0cb980e9
Fix command examples in networking docs

All create commands now need to specify a name for the object to be
created, add this to the "subnet create" examples.

Also, the support for setting the external gateway for a router is still
under review [1], revert to using the old version until the new one is
released.

[1] https://review.openstack.org/357973

Change-Id: Ic97733897d4c158bc4432f861e654a47fef6df60
Closes-Bug: 1648764