Updated openstack/openstack Project: openstack/neutron  d51ac22dd3e42e1fbb3f7778ac321acb70509de6
Check for unbound ports in L3 RPC handler

The handler was making the incorrect assumption that once a
host_id was set, a port which failed to bind could only be in
the 'binding_failed' state. So it would not try to rebind ports
that encountered an exeption during port binding commit that left
them in the unbound state.

Change-Id: I28bbeda5fed4275ea38e27308518f89df9ab4eff
Closes-Bug: #1648879