Updated openstack/openstack Project: openstack/keystone  94fe1773b890b6a368152b88205262328bc6faf5
Remove exception from v2 validation path

In the process of refactoring most of the token provider API, I
proposed a patch that would raise an exception if a v2.0 token was
validated to prove nothing in keystone was currently validating v2.0
tokens. This commit removes the exception from the v2.0 validation
path because we will need to be able to validate v2.0 token during
and upgrade from Newton to Ocata.

The shim should remain in the token provider through-out the Ocata
release, but it can be removed as soon as Pike opens for development.
When deployments upgrade from Ocata to Pike, hopefully deployments
will no longer have any v2.0 tokens and they will all be considered
expired anyway.

Change-Id: I1e50aadee45e16b7bad64de446a6e66c6ad237da