Updated openstack/openstack Project: openstack/nova  909410cdfbbaa796a9ec719ce9e6a7be5f54c2a2
Fix crashing during guest config with pci_devices=None

The Instance.pci_devices field is nullable, but the get_instance_pci_devs()
function does not account for that. If it is passed an instance (from an
older deployment) with no pci_devices, it will crash trying to iterate
None as a list.

Change-Id: I3d535e01ac31db7804347c3938c0d88a28ba67f5
Closes-Bug: #1648887