Updated openstack/openstack Project: openstack/trove  80e7d569a0775b7bf1e7f57f660af0112fa4bd36
Stop caching client to fix long-running test fails

If a node is really slow then a single test can take longer than an
hour. It will then fail as the client it has cached will now have an
expired token. To alleviate this, the caching of the clients was removed
(since the time saved on caching the clients is probably negligible
anyway).

Checking the client code was also cleaned up for a number of tests.

The cluster tests were modified, however the code was reverted as
there is no way to test it at present (they will not run as they
haven't been modified to work with the new neutron set up).  This
will be fixed in a separate changereq (since they are currently not
being tested in the gate, this should not hold up development).

Change-Id: I63d5b78084486b806bff38068260e42128c0d0ed
Closes-Bug: 1647002