Updated openstack/openstack Project: openstack/trove  8adfb7e4f27fb9daf87dda018094cc6c58089b88
Clean up trovestack utility (fix kick-start)

With the recent changes to devstack and the migration
of the trove utility 'redstack' to 'trovestack' a number
of unnecessary items were brought over.  This change
attempts to fix some problems and clean up some issues:

 - Environment variables are ignored.
 - Credentials are taken from the clouds.yaml file or
   sourced from the devstack/accrc/{project}/{user} directory.
   This means that authentication is now done through the
   mechanism specified by devstack.
   This should fix the kick-start and clean issues.
 - The openstack client is used wherever possible with the
   same set of credentials as mentioned above.
 - References to stable/<branch> have been removed.
 - The trove devstack plugin is assumed to be in the trove repo.
 - Sourcing /tmp/integration/env.rc has been removed (items in
   this file can be moved to ~/.trovestack.options.rc if needed).
 - set_home_dir was removed.
 - The fault flavors no longer set up 'resize' versions of themselves.
 - Help for int-tests was improved to add scenario examples and
   a few other minor clarifications.

This should fix the issues people may have been having with
'kick-start' giving an authentication error after sourcing the
openrc file or 'clean' not running properly.  There should be
no more 'plugin already enabled' style errors either.

Note: The original code for switching to V3 of authentication was
removed as it wasn't an issue with kick-start/clean up per se.
It still should be fixed, but as a separate set of changes.

Change-Id: Ib9af99c37f6729e4e91114a8032732d23a3e0850