Updated openstack/openstack Project: openstack/nova  1af093d3e11ef47986a5fafdf8bf09ec358b8267
neutron: handle no_allocate in create_pci_requests_for_sriov_ports

If the server create request is specifically requesting that
networks aren't allocated for the instance, then we should
short-circuit on that in create_pci_requests_for_sriov_ports.

In the case of no_allocate, the NetworkRequestList is not empty,
it has a single NetworkRequest entry, but it's the 'none' value
which is treated special.

Change-Id: I740c49111bc8e5e76a2fe42b35e8d6c9267e04a2