Updated openstack/openstack Project: openstack/nova  0e150163e790b8f07062e0b6d94b200a17340ec5
Make xenapi tests work with CONF.use_neutron=True by default

The main issue in the xenapi unit tests was faking out network_info
to be an unrealistic string, it must be a list of dicts. There was
also a neutron-specific test which wasn't actually running against
neutron before this, so this fixes that test to behave properly.

Part of blueprint use-neutron-by-default

Change-Id: I3a817ada790b9aa827c7844a2e249dd061ed3936