Updated openstack/openstack Project: openstack/nova  5c7a5f937be389aa951c3c482d8f89d3bde5ec64
Improve the error message for failed RC deletion

When an attempt is made to delete a resource class that is used for
existing Inventory, an exception is raised. This change adds the name of
the resource class to the error message, making it a little more
informative to the user.

Blueprint: custom-resource-classes

Change-Id: Id482ccaec3b1916780782164aa3c9e84a8521177