Updated openstack/openstack Project: openstack/nova  cf7a3f0a34574e0b21d6e35cd90815c2ce8010a2
Quiet unicode warnings in functional test_resource_provider

The 'name' filter values passed to the ResourceProviderList object
are expected to be unicode. When not unicode the tests produce
warnings which are noise since that's not what is actually being
tested. This change makes the values unicode and thus shuts up the
warnings.

Change-Id: Iebc50ff2ced120d5004fa6d053401469097beb1b