Updated openstack/openstack Project: openstack/zaqar  110b53bb52e499fd4f382a0ba414e2faad5e0954
Convert user getting started guide to rst

Currently, Zaqar is holding the user related guide as xml format,
which is hard to maintain. Until now, we have moved the config
ref to rst and the API ref is happening as well. So it would be
nice if the user getting started guide can be coverted to rst
as well. This patch does it and removes the old docbook files.

Change-Id: Ib74fb49ba2a5b2c1dea9c6bd75406d804504d392