Updated openstack/openstack Project: openstack-infra/project-config  64a24ea0b547812dea57126a7cda093dab8cfd9a
Reduce min-ready for ubuntu-trusty to 1

Now that our migration to ubuntu-xenial for jobs is finished, we can
now reduce ubuntu-trusty to 1.  I know this is a little agressive, but
since our time to ready for nodes is 30seconds on most clouds, its
something I wouldn't mind test driving to see how nodepool reacts to
the demand.

Change-Id: I3d0cdbba06a265729670ada6c541b46838f448a5
Signed-off-by: Paul Belanger <pabelanger@redhat.com>