Updated openstack/openstack Project: openstack/swift  54d1e1bae8d7a3ff013e7eb4398bb6dca04d6ca6
tempurls with a prefix-based scope

The middleware now allows the usage of signatures with a prefix-based
scope. A prefix-based signature grants access to all objects which share
the same prefix. This avoids the creation of a large amount of signatures,
when a whole container or pseudofolder is shared.

Please see spec: https://review.openstack.org/#/c/199607/

Change-Id: I03b68eb74dae6196b5e63e711ef642ff7d2cfdc9