Updated openstack/openstack Project: openstack/oslosphinx  8d1cf4f0d8df7e6a58ae39c38d38ba9f59690468
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove oslosphinx.

Change-Id: I3daafcf22e2d9a1cd73d389debe3b27eccc91910