Updated openstack/openstack Project: openstack/oslotest  4fbb2c86e34b578e0d09931330161dccd192397d
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove oslotest.

Change-Id: I21e2d0555334a1bd7ad8aa37796f599292b30ed0