Updated openstack/openstack Project: openstack/pycadf  b4bc985324061262c9c5f7f338691376e29a49a3
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove pycadf.

Change-Id: I385d89d2ce2ecc85532f4acc31ff1c975a150178