Updated openstack/openstack Project: openstack-infra/project-config  fb56a3b653fff85112e551b98279dcb2a90d72b2
Tell shade which services are running

Shade has traps in its tests to only run tests on services that exist.
However, that can allow skipping of tests that _should_ run in a given
config. Add explicit variables for the things we expect to exist in the
gate so that if they're missing all of a sudden we'll see test jobs
fail.

Change-Id: I2433c7bced6c8ca785634056de45ddf624031509