Updated openstack/openstack Project: openstack/heat  1190b1bead74761e2e45a265fbdda51148ce806a
Handle ambiguous physical resource IDs

It's possible that we could end up with multiple resources with the same
physical resource ID, but that would be undetectable since we return only
one from the database layer. This change allows us to detect the problem an
return an error where the result is rendered ambiguous.

Change-Id: I2c5ddbe6731c33a09ec7c4a7b91dcfe414da4385