Updated openstack/openstack Project: openstack/tripleo-specs  c20fafe37d2452a23874541f33f3bd71357ef7dc
Spec review: additional note on when to +2

There was a question during the Nov 29 weekly meeting [1] about when is
it ok to +2 a spec, when the other parts do not relate to one's area of
expertise but also to others. This includes the answer proposed into
the policy.

Also, change 'devtest' to 'undercloud install' to reflect the tooling
update.

[1] http://eavesdrop.openstack.org/meetings/tripleo/2016/tripleo.2016-11-29-14.00.log.html

Change-Id: I12c4cdd6c61188a570b8d891d30af36bac794157