Updated openstack/openstack Project: openstack/python-novaclient  883f918dee55bd3c1575aa74f613df44473f50a7
Check source_type and destination_type when booting with bdm provided

When booting instance with block_device_mapping provided,
in the current implementation, the "destination_type" is
allowed to be None, and this lead to un-sync between Nova
and Cinder: Nova will have a bdm in its record while the
volume is still in "available" status in cinder.

Also we should check source_type here

This patch adds a check for source_type and
destination_type, source_type shoud be within
'volume', 'image', 'snapshot' or 'blank';
When empty destination_type is provided, it
will be assigned to an default value according
to source_type, if it is provided with a wrong
value, command error will raise and information
will shown.

Change-Id: If6c7d8ed58a94d262ac6c300e75aca4cdf99ff1d
partial-bug: #1644725