Updated openstack/openstack Project: openstack/cinder  12eccd10f4227582885a25cbbae169a5cb4bd2ab
Do not manage VG with > 1 volume in Kaminario driver

Currently, if we want to manage a volume in the K2 array,
we rename the volume and associated VG in K2
according to cv-CinderVolumeId for volume and
cvg-CinderVolumeId for VG.

So, if K2 volume is managed by cinder, then for any future
reference for the managed volume, cv-CinderVolumeId and
cvg-CinderVolumeId must exist on K2 array.

But if a VG in K2 have more than one volume and we try to
manage another volume in the same VG, then it will rename
the VG and any future reference for another managed volume
will result in error.

Change-Id: I0db48d11d812ea802d11165a959ff5b46a4d6af3
Closes-Bug: #1646766
Co-Authored-By: Nikesh Mahalka <Nikesh.Mahalka.ctr@kaminario.com>
Co-Authored-By: Ido Benda<Ido.Benda@kaminario.com>