Updated openstack/openstack Project: openstack/heat  d3cf51a10a55ae16100e8d4c0f952e6edd7b0a47
Add DeployedServer resource

The DeployedServer resource is a subclass of the BaseServer base class.
It allows for creating Server resources to associate with
SoftwareDeployment's where those servers were not orchestrated via Nova.

This is a use case for TripleO so that the software configuration part
of the deployment can be orchestrated with Heat on servers that were not
provisioned with Nova/Ironic.

Partially-implements: blueprint split-stack-software-configuration
Change-Id: I07b9a053ecd3ef4411b602bbc6ef985224834cf8