Updated openstack/openstack Project: openstack-infra/shade  ee1d90b0eca107787422c1cdd4ce08a244985ad2
Convert test_create_image_put_v2 to requests_mock

Also, add a test for listing an empty list of images as well as to make
sure that an image dict gets propertly normalized when it comes through
in the request payload.

There are likely some more things we can/should assert in the request
history objects.

Co-Authored-By: Jamie Lennox <jamielennox@gmail.com>
Change-Id: Iede134002cb75cb382d07b32115d56fc50a656c6